Test fixes due to ember-data upgrade
For some reason Ember-Data is bringing back synchronous relationships as a
promise.  This was causing the error that the PromiseArray content needs
"The content property of DS.PromiseArray should be set before modifying it".
To bypass this I marked the hasMany side of the relationship as async to get
the tests to pass.  However I think that this might be bug within Ember-data