AMD and CommonJS module-ify the bower dist
ultimately, I decided to export just FactoryGuy from the module. In the
code that creates the module, I also assign FactoryGuyTestMixin to FactoryGuy.testMixin, and Sequence to FactoryGuy.sequence. This allows
the tests to all pass, and means that the distributed JS file will not
add constants to the global namespace (other than FactoryGuy, right now)

It’s kind of hacky to wrap the existing files in a giant function call
by concatenating the start and end of those calls before and after the
other files. Eventually, I’d like to modify the files in src/ so that
they are all modularized and work with one another without the wrapper.